requirements.txt
django>=4.2
djangorestframework>=3.14
mariadb>=1.1
python-dotenv>=1.0
